Refactored validators for schemas
All of the schemas had their own validators for each field; however, they're always the same, and if an error was found on one, then it required fixing the error on the other ones. This makes it easier to maintain since all of the validators are in one place.

If one schema requires a special validator, then it can be written inline on the schema itself (i.e. if the UTM schema required a special case to validating the orthoHt, then it could just be written on the UTM schema where the validators.orthoHt is referenced).